chore: 더미 데이터 파일 이동
